backend/requirements.txt
# Core FastAPI
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart

# Your RAG Stack
llama-index-core>=0.12.0
llama-index>=0.12.0
llama-index-multi-modal-llms-gemini
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
transformers>=4.41.0
sentence-transformers
google-generativeai

# Video Processing
youtube-transcript-api
yt-dlp
opencv-python

# Database & Storage
sqlalchemy>=2.0.0
alembic

# Utilities
python-dotenv
pydantic>=1.8.0
pydantic-settings
httpx
redis>=5.0.0  # Using redis.asyncio
websockets

# Development
pytest
pytest-asyncio
black